Move top board edge inwards to clear USB-A jack
On version 1.0 the jack doesn't *quite* fit without a little sanding. It turns out that the footprint's silkscreen marker for the top of the board clearance area is off by 0.1 mm (3 mm in the footprint, 2.9 mm in the datasheet), and my board edge followed the grid so it was 0.05 mm past even that. This commit pulls the board edge back to properly accomodate the connector.

Now according to the datasheet the lower edge of the connector is 18.8 mm from the tip, with the mounting tabs 2 mm long. Because the origin of the footprint is the center of these mounting tabs, that makes the origin 17.8 mm offset from the tip. The end of the cutout area is given as 14.90 +/- 0.2 from the tip, which is an offset of 2.9 +/- 0.2 from the footprint origin. Accordingly I set the top board edge to 2.7 mm above the USB footprint's origin to provide clearance up to the max provided tolerance.

Science! ...that I should have probably done the first time around.